pyproject.toml
[project]
name = "bothackathon"
version = "0.1.0"
requires-python = "~=3.13.0"
dependencies = [
    "bunnet>=1.3.0",
    "google-generativeai>=0.8.4",
    "gtts>=2.5.4",
    "pydantic>=2.10.6",
    "pymongo>=4.11.2",
    "pytelegrambotapi>=4.26.0",
]

[dependency-groups]
dev = [
    "ipython>=9.0.0",
]
